Move lookups to repo layer
