+++ app/api/api_v1/routers/summaries.py
@@ -12,7 +12,7 @@ from fastapi import APIRouter, Depends, HTTPException, Request, status
 from app.api.api_v1.schemas.search import GeographySummaryFamilyResponse
 from app.clients.db.session import get_db
 from app.core.browse import BrowseArgs, browse_rds_families
-from app.core.lookups import get_country_slug_from_country_code, is_country_code
+from app.repository.lookups import get_country_slug_from_country_code, is_country_code
 
 _LOGGER = logging.getLogger(__name__)
