+++ app/core/ingestion/pipeline.py
@@ -15,8 +15,8 @@ from db_client.models.organisation import Organisation
 from sqlalchemy.orm import Session
 
 from app.api.api_v1.schemas.document import DocumentParserInput
-from app.core.lookups import doc_type_from_family_document_metadata
 from app.repository.geography import get_geo_subquery
+from app.repository.lookups import doc_type_from_family_document_metadata
 
 _LOGGER = logging.getLogger(__name__)
