+++ tests/search/vespa/test_range_and_keyword_filters_search.py
@@ -5,7 +5,7 @@ from db_client.models.dfce import Geography
 from fastapi import status
 
 from app.api.api_v1.routers import search
-from app.core.lookups import get_country_slug_from_country_code
+from app.repository.lookups import get_country_slug_from_country_code
 from tests.search.vespa.setup_search_tests import (
     VESPA_FIXTURE_COUNT,
     _make_search_request,
